Hi all,

This pull request contains a backport of commit [47df1459](https://github.com/openjdk/jdk/commit/47df14590c003ccb1607ec0edfe999fcf2aebd86) from the [openjdk/jdk](https://git.openjdk.org/jdk) repository.

The commit being backported was authored by Amit Kumar on 10 Apr 2024 and was reviewed by Lutz Schmidt and Martin Doerr.

Thanks!

Benchmarking: 
```
./build/linux-s390x-server-release/jdk/bin/java -Xms4g -Xmx4g -jar dacapo-9.12-MR1-bach.jar h2 -s huge -t 1 -n 1

Without the patch:

===== DaCapo 9.12-MR1 h2 PASSED in 209768 msec ===== 
===== DaCapo 9.12-MR1 h2 PASSED in 197472 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 222872 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 215725 msec =====


with patch: 
===== DaCapo 9.12-MR1 h2 PASSED in 171055 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 169696 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 174032 msec =====
===== DaCapo 9.12-MR1 h2 PASSED in 167271 msec =====
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] [JDK-8310513](https://bugs.openjdk.org/browse/JDK-8310513) needs maintainer approval

### Issue
 * [JDK-8310513](https://bugs.openjdk.org/browse/JDK-8310513): [s390x] Intrinsify recursive ObjectMonitor locking (**Enhancement** - P4 - Approved)


### Reviewers
 * [Lutz Schmidt](https://openjdk.org/census#lucy) (@RealLucy - **Reviewer**) ⚠️ Review applies to [50b25633](https://git.openjdk.org/jdk21u-dev/pull/537/files/50b256332732c27188156ffbc71040851c4b8110)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk21u-dev.git pull/537/head:pull/537` \
`$ git checkout pull/537`

Update a local copy of the PR: \
`$ git checkout pull/537` \
`$ git pull https://git.openjdk.org/jdk21u-dev.git pull/537/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 537`

View PR using the GUI difftool: \
`$ git pr show -t 537`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk21u-dev/pull/537.diff">https://git.openjdk.org/jdk21u-dev/pull/537.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk21u-dev/pull/537#issuecomment-2079734826)