@offamitkumar 2024-04-29
>you can force the state to "clean" by adding the "/clean" tag.

ohh, thanks, I didn't know that it is possible to mark PR forcefully clean.  I had missed it while looking at bot-commands.